Modified monster_01 asset.
Changed the joints connecting the elbows to the upper arms to fixed so
that there is less bouncing and separation of the arms.

Did the same thing with the knees in connecting them fixed joints to the
thighs.

Have 2 red point lights in for the eyes as well.

There is also a low intensity directional light on the monster to
illuminate the meshes.